Commiting for review

+++ generate_text/generate_character.py
@@ -20,7 +20,7 @@ def drawChar(img_size, chars, font_size, font_path, color=0):
     font = ImageFont.truetype(font_path, int(img_size * font_size))
     img_PIL = Image.fromarray(np.full((img_size, img_size), 255, dtype=np.uint8), mode='L')
     draw_PIL = ImageDraw.Draw(img_PIL)
-    coordinates = randomize_location(font, chars, x_range=img_size, y_range=img_size, out_of_bounds_threshold=-img_size/12.5) # Top-left of character
+    coordinates = randomize_location(font, chars, x_range=img_size, y_range=img_size, out_of_bounds_threshold=-int(img_size/12.5)) # Top-left of character
     draw_PIL.text(coordinates, chars, font=font, fill=color)
     return np.array(img_PIL)
